requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
bleach==6.3.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
cryptography==45.0.2
defusedxml==0.7.1
Django==5.0.6
django-allauth==0.59.0
django-browser-reload==1.18.0
django-environ==0.11.2
django-tailwind==3.8.0
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.2.2
packaging==25.0
pillow==10.4.0
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
setuptools==80.8.0
six==1.17.0
sqlparse==0.4.4
stripe==8.0.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
whitenoise==6.6.0
